Point the javadoc at methods that exist, and mend what the inspections found
**Six {@link} targets named methods that are not there.** `MMU#tickDMA` never
existed in this repository's history -- `beginDMACycle` is what copies a page
into OAM and what fetches the DMC's sample, which is exactly the hole both
listeners are documenting. `IPSPatch#read` resolved to the *private* `read(byte[],
int, int)` rather than to the static factory, so the link the reader wanted was
inaccessible; it names its arguments now. `#present(int[])` lost its frame phase
when the decoder gained one. And `#describe` was reaching for `#parts`, whose own
javadoc states the very reason it borrows.

**Four javadoc blocks had come adrift from the methods they describe**, each by
the same accident: a block was written above a method that already had one, and
the older block stayed where it was rather than moving down with its own. So
`setPalette`, `Repl.hack`, `OAMViewerFrame.refresh` and
`emphasisPicksTheDimmedCopyOfTheColour` were all undocumented while their prose
sat one method up, describing something else.

**The rest are small, and each is what it looks like.** Sprite evaluation had a
`case FINISHED` that the guard eight lines above it makes unreachable.
`ScreenComponentTests` dereferenced a `@Nullable` snapshot eleven times, and goes
through a helper now that says which of the two things went wrong. Two assertions
did integer division inside a `double` comparison -- 341 dots to 3 CPU cycles is
not a whole number, so the sum says so now -- and a `byte` was flipped with an
`int`. `Math.clamp` and `StringBuilder.repeat` where they read better.

`Tracer` keeps its padding loop, which is the one suggestion refused for a reason
rather than for taste: `Overclock.MAX_SCANLINES` is 1000 a half, so a four digit
scanline against a width of three hands `repeat` a count of -1 and throws, where
the loop silently skips. `GameGenie.substitute` keeps its indexed loop for the
reason written three lines above it, and `pipelineHasTheBus` keeps its name
rather than inverting three paragraphs of javadoc to save a `!`. The 41 reports
of a parameter always having the same value are nearly all test helpers whose
parameter names are what say what the literal means.

Nothing observable moves. The 211 tests pass, `javadoc -private
-Xdoclint:reference` is clean across all four modules, a rewound machine still
hashes identical to one that never went forward, and a 250 line `--format text`
session differs only in the `host` block.

+++ mynes-core/src/test/java/com/github/dimiro1/mynes/video/NTSCFilterTests.java
@@ -166,7 +166,7 @@ class NTSCFilterTests {
             var pixels = filter.colourise(frame, 0);
             var here = luma(pixels[MIDDLE]);
             var spread = here + luma(pixels[MIDDLE - 1]) + luma(pixels[MIDDLE + 1]);
-            var kept = (double) here / spread;
+            var kept = here / spread;
 
             assertTrue(
                     kept > previous,
